refactor to createBrowserChannel
