+++ code/builders/builder-webpack5/templates/virtualModuleModernEntry.js.handlebars
@@ -1,14 +1,14 @@
 import { global } from '@storybook/global';
 
 import { ClientApi, PreviewWeb, addons, composeConfigs } from '@storybook/preview-api';
-import { createPostMessageChannel } from '@storybook/channels';
+import { createBrowserChannel } from '@storybook/channels';
 
 import { importFn } from './{{storiesFilename}}';
 
 const getProjectAnnotations = () =>
   composeConfigs([{{#each previewAnnotations}}require('{{this}}'),{{/each}}]);
 
-const channel = createPostMessageChannel({ page: 'preview' });
+const channel = createBrowserChannel({ page: 'preview' });
 addons.setChannel(channel);
 
 if (global.CONFIG_TYPE === 'DEVELOPMENT'){
