+++ code/lib/channels/src/index.ts
@@ -2,10 +2,40 @@
 /// <reference path="./typings.d.ts" />
 
 import { Channel } from './main';
+import { PostMessageTransport } from './postmessage';
+import type { ChannelTransport, Config } from './types';
+import { WebsocketTransport } from './websocket';
 
 export * from './main';
 
 export default Channel;
 
-export { createChannel as createPostMessageChannel } from './postmessage';
-export { createChannel as createWebSocketChannel } from './websocket';
+export { createChannel as createPostMessageChannel, PostMessageTransport } from './postmessage';
+export { createChannel as createWebSocketChannel, WebsocketTransport } from './websocket';
+
+type Options = Config & {
+  extraTransports?: ChannelTransport[];
+};
+
+/**
+ * Creates a new browser channel instance.
+ * @param {Options} options - The options object.
+ * @param {Page} options.page - The puppeteer page instance.
+ * @param {ChannelTransport[]} [options.extraTransports=[]] - An optional array of extra channel transports.
+ * @returns {Channel} - The new channel instance.
+ */
+export function createBrowserChannel({ page, extraTransports = [] }: Options): Channel {
+  const transports: ChannelTransport[] = [new PostMessageTransport({ page }), ...extraTransports];
+
+  if (CONFIG_TYPE === 'DEVELOPMENT') {
+    const protocol = window.location.protocol === 'http:' ? 'ws' : 'wss';
+    const { hostname, port } = window.location;
+    const channelUrl = `${protocol}://${hostname}:${port}/storybook-server-channel`;
+
+    transports.push(new WebsocketTransport({ url: channelUrl, onError: () => {} }));
+  }
+
+  return new Channel({ transports });
+}
+
+export type { Listener, ChannelEvent, ChannelTransport, ChannelHandler } from './types';
