+++ code/lib/preview-api/src/modules/preview-web/PreviewWeb.integration.test.ts
@@ -26,7 +26,7 @@ import {
 
 jest.mock('@storybook/channels', () => ({
   ...jest.requireActual('@storybook/channels'),
-  createPostMessageChannel: () => mockChannel,
+  createBrowserChannel: () => mockChannel,
 }));
 jest.mock('@storybook/client-logger');
