+++ code/ui/manager/src/globals/exports.ts
@@ -113,7 +113,14 @@ export default {
     'resetComponents',
     'withReset',
   ],
-  '@storybook/channels': ['Channel', 'createPostMessageChannel', 'createWebSocketChannel'],
+  '@storybook/channels': [
+    'Channel',
+    'PostMessageTransport',
+    'WebsocketTransport',
+    'createBrowserChannel',
+    'createPostMessageChannel',
+    'createWebSocketChannel',
+  ],
   '@storybook/core-events': [
     'CHANNEL_CREATED',
     'CONFIG_ERROR',
